feat: Throw Error when trying to delete a category with some Art. Save category name in snake_case

+++ backend/src/modules/Categories/useCases/DeleteCategory/DeleteCategoryUseCase.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from 'tsyringe';
+import { prisma } from '../../../../services/prisma';
 import { ICategoryRepository } from '../../repositories/ICategoryRepository';
 
 @injectable()
@@ -9,6 +10,19 @@ export class DeleteCategoryUseCase {
     ) { }
 
     async execute(id: string): Promise<void> {
+        const category = await prisma.category.findUnique({
+            where: {
+                id
+            },
+            include: {
+                Art: true
+            }
+        })
+
+        if (category.Art.length !== 0) {
+            throw new Error(`Cannot delete this category, there ${category.Art.length === 1 ? 'is' : 'are'} ${category.Art.length} ${category.Art.length === 1 ? 'Art' : 'Arts'} with this category, delete them before.`)
+        }
+
         await this.categoriesRepository.delete(id)
     }
 }
